<!-- Thanks for sending a pull request!
1. If this is your first time, please read the [contributing guide](https://metallb.universe.tf/community/#contributing)
2. For non-trivial pull requests, please [file an
   issue](https://github.com/metallb/metallb/issues/new) first, and get
   agreement that the change is a good idea, and a general guideline
   for how it should be implemented, before sending code. Large PRs
   that weren't first discussed and agreed upon in an issue won't be
   accepted.
3. If the PR fixes a particular bug, please include the words "Fixed
   #<issue number>" in the PR text, so that the bug auto-closes when
   the PR is merged.
-->

**Is this a BUG FIX or a FEATURE ?**:

> Uncomment only one, leave it on its own line:
>
> /kind bug
/kind cleanup
> /kind feature
> /kind design
> /kind flake
> /kind failing
> /kind documentation
> /kind regression

**What this PR does / why we need it**:

  metav1.LabelSelectorAsSelector is the biggest offender in terms of
  performances in the validation function. Here we add an lru cache so the
  result can be cached for reasonable numbers.
  
  Also, avoid deepcopy as it puts more stress on the GC.


Before

```
go test -bench=BenchmarkValidateConfig -benchmem -benchtime=10s -count=3
goos: linux
goarch: amd64
pkg: github.com/metallb/frr-k8s/internal/webhooks
cpu: Intel(R) Xeon(R) Gold 6230R CPU @ 2.10GHz
BenchmarkValidateConfig-104                 4042           2949891 ns/op          959868 B/op      23363 allocs/op
BenchmarkValidateConfig-104                 4044           2960876 ns/op          960273 B/op      23363 allocs/op
BenchmarkValidateConfig-104                 4033           2956161 ns/op          960348 B/op      23363 allocs/op
PASS
ok      github.com/metallb/frr-k8s/internal/webhooks    35.852s
```
After

```
go test -bench=BenchmarkValidateConfig -benchmem -benchtime=10s -count=3
goos: linux
goarch: amd64
pkg: github.com/metallb/frr-k8s/internal/webhooks
cpu: Intel(R) Xeon(R) Gold 6230R CPU @ 2.10GHz
BenchmarkValidateConfig-104                 7335           1621307 ns/op          275721 B/op       9227 allocs/op
BenchmarkValidateConfig-104                 7335           1624197 ns/op          275670 B/op       9227 allocs/op
BenchmarkValidateConfig-104                 7347           1625323 ns/op          275672 B/op       9227 allocs/op
```

**Special notes for your reviewer**:

**Release note**:
<!--  Write your release note:
1. Enter your extended release note in the below block. If the PR requires additional action from users switching to the new release, include the string "action required".
2. Follow the instructions for writing a release note from k8s: https://git.k8s.io/community/contributors/guide/release-notes.md
3. If no release note is required, just write "NONE".
-->

```release-note
Webhook performance improvements on scale
```